add more documentation to readme and command system
